refactor(Worklets): remove public header dir
