+++ docs/docs-worklets/docs/threading/scheduleOnRuntime.mdx
@@ -48,7 +48,7 @@ Arguments to the function you want to execute on the [Worker Runtime](/docs/fund
 ## Remarks
 
 - The worklet is scheduled on the Worker Runtime's [Async
-  Queue](https://github.com/software-mansion/react-native-reanimated/blob/main/packages/react-native-worklets/Common/cpp/worklets/Public/AsyncQueue.h)
+  Queue](https://github.com/software-mansion/react-native-reanimated/blob/main/packages/react-native-worklets/Common/cpp/worklets/RunLoop/AsyncQueue.h)
 - The worklet cannot be scheduled on the Worker Runtime from [UI
   Runtime](/docs/fundamentals/runtimeKinds#ui-runtime) or another [Worker
   Runtime](/docs/fundamentals/runtimeKinds#worker-runtime), unless the [Bundle Mode](/docs/experimental/bundleMode) is enabled.
