+++ docs/docs-worklets/src/theme/TOCItems/index.js
@@ -1,6 +1,5 @@
 import { TOCItems } from '@swmansion/t-rex-ui';
 
-
 export default function TOCItemsWrapper(props) {
   return (
     <>
