+++ packages/react-native-worklets/src/runtimes.native.ts
@@ -115,7 +115,7 @@ export function createWorkletRuntime(
  * for more information about the different runtime kinds.
  *
  * - The worklet is scheduled on the Worker Runtime's [Async
- *   Queue](https://github.com/software-mansion/react-native-reanimated/blob/main/packages/react-native-worklets/Common/cpp/worklets/Public/AsyncQueue.h)
+ *   Queue](https://github.com/software-mansion/react-native-reanimated/blob/main/packages/react-native-worklets/Common/cpp/worklets/RunLoop/AsyncQueue.h)
  * - The function cannot be scheduled on the Worker Runtime from [UI
  *   Runtime](https://docs.swmansion.com/react-native-worklets/docs/fundamentals/runtimeKinds#ui-runtime)
  *   or another [Worker
